requirements.txt
playwright
pytest
python-dotenv
pydantic
openai
pillow
